log: Fix file being reported by connection log
Since the introduction of connection logging the log messages always
contain the file name of the logger. For example:

  connection.go:23: [0] connection from 127.0.0.1:52776
  connection.go:23: [0] Account login: 0f8ec4999db083f29369d93731c837f4.wrj
  connection.go:23: [0] frontend closed
  connection.go:23: [0] connection closed

This is uninformative and pointless. However, but using log.Output
instead of log.Printf we can specify a number of frames to skip when
calculating the file name. By using a call depth of 2 we can get the
caller of the function:

  client.go:80: [0] connection from 127.0.0.1:52778
  login.go:173: [0] Account login: 0f8ec4999db083f29369d93731c837f4.wrj
  client.go:331: [0] frontend closed
  client.go:346: [0] connection closed

This is more informative and useful.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>