comms: Fix write timeout for client connection
Currently the read and write timeouts for a client connection are set to
the Server.IdleTimeout value from the configuration file. While we want
the read timeout to be a large value - a player me be idle watching
other players or NPCs or waiting for someone else to login so 10, 30 or
60 minutes could be reasonable. For writing data it should go straight
to the client to be immediately displayed, we should not be waiting for
for the 10, 30 or 60 minutes!

This change sets the write timeout to 10 seconds which should be long
enough, even on a really slow connection with a really slow machine, as
we are only sending low volumes of data[1]. The timeout is currently
defined as a constant, to be moved into the configuration file at a
later date if required.

[1] A complete 80x24 ASCII screen is 1920 characters, just under 2kb.
    More if using colours and Unicode - a full Unicode display using
    only 4 byte characters would be 8kb. If using a lot of colours maybe
    add another 2kb. So the minimum client upload line speed in theory
    would be 1kb/sec before we start timing out.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>