comms: Add a new Ring type + tests implementing a ring buffer
Ring is a very specific implementation of a ring buffer with a fixed
capacity for use in connection limiting. The implementation is a power
of 2 ring buffer with free running indexes and bit-masking instead of
modulo arithmetic. Copious comments have been provided to make tweaking
and repurposing the ring buffer easier.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>