recordjar: Don't identify fields within free text sections
Within a free text section fields should not be identified as such, they
should be treated as part of the free text.

For example:

  %%
    Ref: note
   Name: a note
  Alias: note

  Written on the note is a cryptic message:

     CODE: xyzzy

  What can it mean?
  %%

In this example we should not have a field named 'CODE' with a value of
'xyzzy' as it is within the free text section.

Tests added for fields within free text sections. Also added additional
tests for comment within free text sections which should also be part of
the free text.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>