recordjar: Simplify comment and code for free text 'joiner'
When building up the free text section the joiner for the lines can be
either a line feed '\n' or a space depending on the preceding text and
current line.

This commit adds a new startWS flag that is true if the current line
started with whitespace before any trimming. This is then used to
simplify the logic for the joiner. Comment for the joiner logic also
rewritten.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>