record: Don't reorder StringList when encoding/decoding
Currently when a StringList is encoded or decoded the resulting strings
are sorted. However there is no way of knowing the original order of the
strings, which may be important.

This commit removes sorting from the decoder.StringList and
encoder.StringList functions. Tests updated to check for original
ordering instead of sorted ordering.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>