recordjar: Ignore indented record separator within free text
Currently a record separator '%%' within a free text section will be
identified as a record separator, splitting the free text section into
multiple records. This is irrespective of whether the record separator
has leading or trailing whitespace or not.

With this change a record separator after a free text section must not
have any leading whitespace. A record separator within a free text
section with leading whitespace will be treated as part of the free text
section text. Furthermore, if a record separator appears within a free
text section with no leading whitespace it will be identified as an
actual record separator and the start of a new record.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>