recordjar: Preserve line feeds in free text sections when reading jars
When reading a record jar only significant whitespace is currently
preserved. Significant whitespace includes blank lines and indenting
spaces and/or tabs. This can be inappropriate in some instances and a
wasted effort in others - all text is formatted through the text.Fold
function anyway.

This commit updates the recordjar.Read function to preserve line feeds
'as is'. This simplifies the code to handle free text as we no longer
have to track whether the last line was empty or not.

All tests have been updated where the free text section would have been
effected.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>