recordjar: Fix broken tests due to commit 4c6528f
Commit 4c6528f 'recordjar: Preserve line feeds in free text sections
when reading jars' broke the read tests by including an additional flag
to the recordjar.Read function. However the additional boolean flag was
not implemented in the end. The flag was intended to be preserveLF -
preserve line feed true/false. Instead recordjar.Read was modified to
always preserve line feeds making the flag redundant.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>