attr,has,cmd: Treat qualifiers separate from aliases in attr.Alias
At the moment qualifiers are treated the same as aliases. This leads to
the situation where all of these commands would be valid:

  WHICH BALL
  WHICH GREEN BALL
  WHICH +GREEN

However qualifiers should only apply when used with an alias. They
should not be usable directly as in "WHICH +GREEN".

This change updates attr.Alias so that qualifiers are stored separately
from aliases. It is then possible to check for an alias or specifically
for a qualifier.

New methods added for qualifiers are: HasQualifier and Qualifiers. These
have also been added to the has.Alias interface.

The methods NewAlias, Copy, Marshal, and Dump have been updated to
handle qualifiers being stored separately. Dump will now list aliases
and qualifiers. A helper dumpFmtList has been added to format the
aliases and qualifiers.

The previous TODO in aliases.go to add prefixes has been dropped now we
have qualifiers. A BUG has been noted as we cannot have qualifiers for
specific aliases.

A commented out, failing test in which_test.go has been uncommented as
it now passes. The corresponding BUG has been dropped.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>