attr,has,cmd: Add support for bound qualifiers
Assume the following aliases for an item:

  Aliases: +SHORT SWORD SHORTSWORD

Currently it is possible to use the following commands:

  WHICH +SHORT SWORD
  WHICH +SHORT SHORTSWORD

The second command is undesirable as the qualifier 'SHORT' should only
be usable with the alias 'SWORD' and not 'SHORTSWORD'.

This change introduces bound qualifiers, that is qualifiers that can be
bound to a specific alias. The format is '+qualifier:alias'. If the
alias is not already specified it will be added. For example both of
these definitions are equivalent:

  Aliases: +SHORT:SWORD SWORD SHORTSWORD
  Aliases: +SHORT:SWORD SHORTSWORD

The alias '+SHORT:SWORD' binds the qualifier 'SHORT' to the alias
'SWORD'. This means that:

  WHICH SHORT SWORD

is valid, while:

  WHICH SHORT SHORTSWORD

is not because 'SHORT' is not bound to 'SHORTSWORD' only to 'SWORD'.
If we amend the definition to:

  Aliases: +WOODEN +SHORT:SWORD SWORD SHORTSWORD

The qualifier 'WOODEN' is unbound and so the following are all valid:

  WHICH WOODEN SWORD
  WHICH WOODEN SHORTSWORD
  WHICH WOODEN SHORT SWORD
  WHICH SHORT WOODEN SWORD

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>