attr: reimplement Inventory using double linked list
Currently Inventory are implemented using []has.Thing for contents and
disabled items. This incurs a lot of bookkeeping adding and removing
items and when trying to prevent the capacity of the slices from growing
due to a sudden influx of items which are then removed. This is seen a
lot with locations when a number of players move into a location and
then leave - the slices grow to accommodate the players and then have to
be shrunk else space is wasted.

This change implements a doubly linked list optimised for Inventory
has.Thing items. For the common case of removing an item from one
Inventory and adding it to another - moving location to location,
get/drop item - there is a move method which will reused the node from
the source list for the destination list.

The current list implementation has the overhead of a sentinel node for
the head and the tail of the list for simplicity.

Using a doubly linked list reduces the bookkeeping complexity of
Inventories, only uses the space required and reduces the number of
allocations significantly.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>