attr,has,cmd: Rename Inventory.Players to Inventory.Occupied
Currently we have the methods Narratives, Contents, Disabled and
Everything which return a []has.Thing of the appropriate items. Then we
have Players which returns a bool and not a []has.Thing of players in
the Inventory.

This change renames Players to Occupied so that a Players can be
implmeneted that returns a []has.Thing of players in the Inventory.

Comment also updated in cmd/say.go

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>