event: Add a small timer cache
Creating a new time.Timer is quite expensive. This change adds a small
Timer cache in the form of a buffered channel. The cache only needs to
be big enough to hold a Timer while an event it being processed. The
rest of the time the Timer will be 'in flight'. The typical cycle is:
Timer fires and is cached, event processed, event requeued using
previously cached Timer.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>