cmd: Prevent GET commands on a Player
Yes, it's fun to pick up another player and put them in the iron bound
chest or into the pond or the crack in the wall by the shed :) However,
it also causes a logistical nightmare - you pick up a player and quit,
then the mess has to be sorted out, a bigger mess if the player was put
into a container you are carrying :(

This commit adds a check to the GET command to disallow getting another
Player. If you cannot GET another Player you also can't PUT them into a
container. This also means that when handling items an inventory.Contents
call instead of calling inventory.Players+inventory.Contents is enough.
This simplifies code and reduces corner cases.

In theory it should have been possible to add a VETO to the Player type,
as has been done for the JUNK command. However, due to the way the GET
command checks for vetoes, that ends up vetoing the GET of anything in a
location.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>