text: Fix final line feed turned into space in Unfold + simplify
Currently The Unfold function is turning the final line feed in a block
of text in to a space - with the expectation of joining to the next,
non-existent, line.

This change fixes Unfold and also simplifies the code by adding two
sentinel line feeds - one at the end of the input data and one at the
start of the output data. Both are removed from the final output.

All current tests pass with additional corner cases found during testing
added.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>