comms: Add a quota.TimeSource that can be mocked+update testing
Currently the tests for quota use time.Sleep and as a consequence take
over 30 seconds to run.

This change adds a quota.TimeSource type that in normal use is the
time.Now function. A mock implementation of a quota.TimeSource can now
be provided for testing.

The quota tests have been updated with a mock implementation of
quota.TimeSource that can be controlled with non-second precision and
allows for time to be fast forwarded, reversed or frozen easily. All
instances of time.Sleep have been changed to use the mocked TimeSource
which allows the tests to run at full speed.

Quota tests now complete in under a second.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>