cmd: Fix matcher switching from not enough to unknown
When matching aliases there is currently a check that creates a new
result if switching from finding an item to finding an unknown match.
However we do not check if switching from a 'not enough' match to an
'unknown' match. In this case it is possible to have a result that flags
a 'not enough' match and an 'unknown' match in the same result.

This results in the 'not enough' being shadowed and ignored. For
example, assume we have some chalk and a single apple:

  >WHICH CHALK BALL 2ND APPLE
  You see some chalk here.
  You see no 'BALL' here.
  >

There should also be a "You don't see that many ‘APPLE’ here." message.

This change adds a check for a 'not enough' match before switching to an
'unknown' match.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>