cmd: Add matcher tests to replace WHICH command tests
At the moment there are tests written for the WHICH command. However the
tests do not test the WHICH command itself but instead exercise the
matcher code.

This change recreates all of the current WHICH tests as equivalent
matcher tests. In addition there are some new tests that cover some
recently found issues with the marcher.

The next step will be to amend the WHICH tests to actually test the
WHICH command.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>