cmd: Matcher should only return unique Thing in results
It is intended that any single call of Match, MatchLimit or MatchAll
return only unique Thing. Note that this does not hold for consecutive
calls, such as Match followed by MatchAll.

However, the current check for equality incorrectly compares a Result
type with a Thing type - which will never match. The correct check
should be comparing Result.Thing and Thing..

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>