cmd: Update WHICH tests to test command and not matcher
The current tests test the functionality of the matcher and not the
WHICH command itself. For the WHICH command we are interested in finding
items in the correct inventories and the colourisation of the messages
sent to players. This change updates the tests to test the WHICH command
itself and not the matcher.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>