cmd: Improve performance of specialQualifier/lastLeadingDigit
Currently the function lastLeadingDigit only returns the position of the
last leading digit in a string. The specialQualifier function then
slices the string based on the returned value and passes it to
strings.Atoi to convert the digits into an integer.

This commit reimplements lastLeadingDigit as leadingDigits. The new
function returns the digits as an integer and returns a count of the
digits used. Using strings.Atoi is no longer required. Example:

  leadingDigits("123xyz") // n=123, count=3

The specialQualifier function has been reworked to use the new
leadingDigits. Tests and benchmarks have also been updated. On average
specialQualifier is now twice as fast. Benchmarks:

  benchmark                              old ns/op  new ns/op  delta
  BenchmarkSpecialQualifier/ALL-4        3.86       3.56       -7.77%
  BenchmarkSpecialQualifier/1-4          15.4       8.03       -47.86%
  BenchmarkSpecialQualifier/12-4         17.5       8.91       -49.09%
  BenchmarkSpecialQualifier/123456-4     25.8       12.6       -51.16%
  BenchmarkSpecialQualifier/1-2-4        27.2       13.6       -50.00%
  BenchmarkSpecialQualifier/12-34-4      33.1       16.0       -51.66%
  BenchmarkSpecialQualifier/1234-5678-4  39.1       20.4       -47.83%
  BenchmarkSpecialQualifier/SHORT-4      7.44       6.81       -8.47%

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>