cmd: Overhaul Match and MatchAll, drop MatchLimits
Comments have been made that the current matcher code is an unwieldy
mess :( I agree :)

This commit restructures and rewrites a lot of the functionality to be
more readable and easier to maintain. In particular results are now
subsetted in separate passes: subsetAlias→subsetQualifiers→subsetLimits

All current tests are passing. New tests to check consecutive 'unknown'
results are merged, but consecutive 'not enough' results are not.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>