attr: Add Pending method to action, cleanup and reset events
At the moment there is no coherency between the action, cleanup and
reset events. Really there should be an Event interface that they
implement.

For now this change implements a Pending method which allows the use of
an interface{ Pending() bool } to ease some testing. However, I'm not
sure if that's a good idea even - if you call Pending and then act on it
the event could have fired immediately after calling Pending but before
action was taken. At best it has limited value and will need reviewing
when the events are unified under a common interface.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>