attr: Veto 'PUT <item> <player>' via PUTIN pseudo command by default
At the moment it is possible to put an item 'into' a player. This is
because players have an Inventory and so are seen as being containers.
This change adds a Veto to Player for the PUTIN pseudo command to
prevent this. At a later stage a GIVE command will be added which will
allow players to give items to other players if the receiving player
consents.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>