cmd: Check TAKEOUT veto before checking item in container
Currently the TAKE command checks if the item is in a container before
checking if the TAKE command is going to be vetoed. This results in a
player being able to fish for items. For example, a bag with a TAKEOUT
veto of "You can't get the bag open.":

  >GET BALL BAG
  The bag does not contain a 'BALL'.
  >

Now the player knows there is no ball in the bag. However if there was a
ball in the bag the player would see:

  >GET BALL BAG
  You can't get the bag open.
  >

In both cases the player should see the "You can't get the bag open."
message. This change updates the order in which the TAKEOUT veto is
checked so that it is always checked before the items in the container.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>