config,docs: Fix tests for Go1.13 (drop cli args for env var)
The Go 1.13 release notes[1] contain the following statement:

 "Testing flags are now registered in the new Init function, which is
 invoked by the generated main function for the test. As a result,
 testing flags are now only registered when running a test binary, and
 packages that call flag.Parse during package initialization may cause
 tests to fail."

The above change and the fact that config.init indirectly calls
flags.Parse causes the tests for WolfMUD to fail with:

 flag provided but not defined: -test.testlogfile

Go issue #31859 is also relevant to this problem.

This change drops the usage of command line arguments altogether to
specify the configuration file to use. Instead the server now uses the
WOLFMUD_DIR environment variable if it is set. This change also updates
the relevant WolfMUD documentation.

 [1] https://golang.org/doc/go1.13
 [2] https://github.com/golang/go/issues/31859

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>