cmd: Drop impossible check in findContainer for TAKE cmd
For the TAKE command we have to find a valid container before we can
look inside it and identify any items. Therefore if we find something
for the container we should not treat it as an item at any point.

This makes the test 'noItem && !inv.Found()' as nonsense as we can never
say 'What did you want to take XXX from?' as we can never know the item
and not treat it as a container.

For example if we have 'TAKE BALL' we either have a ball, which we can
find, or we wanted to take the ball from a container and so cannot find
the container as it is not specified. In the first case we would say
'You cannot take anuthing from the ball.' in the second case we would
say 'You see no ball to take things out of.'

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>