cmd: Improve EXAMINE messages
Use TheName for actor messages. Make the 'You examine' text good and
start the actual description on a new line. Add 'they are carrying' to
observer messages otherwise the actor could examine something they are
holding and the observer could then be looking around for the item,
wondering where it is.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>