cmd: Fix nil pointer panic in JUNK
It is possible for s.where to be nil if the JUNK is happening as the
result of a CLEANUP. In which case the actor is the location itself in
which case the location has no 'where' to be. Then this happends the
value of s.where is nil and calling s.where.Everything panics. If
s.where was {*Inventory)(nil) instead it would be okay. However this
requires a bigger fix to attr.Locate and all usage of where throughout
the code :(

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>