attr,has: Add a Thing.Freed method
Currently 'len(Thing.Attrs()) == 0' is used to test if a thing has been
freed. However calling Thing.Attrs causes a new Thing.attrs to be
allocated and copied which is wasteful as the allocated copy is then
thrown away.

This change adds a Thing.Freed method which can use len() directly on
the Thing's attribute slice without allocation or copying.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>