attr,has: Add Attribute.Is to compare for specific types
This change adds an Is method to attr.Attribute type and the
has.Attribute interface. It's purpose is to return true if the passed
attribute implements a specific interface - usually the interface for
the attribute type. For example attr.Name would test for the has.Name
interface. The name Is was chosen to read 'item.Is(interface)' and
return true or false.

For development only an attr.Attribute.Is method has been addedi, that
always rerturns false, until all attributes implement Is themselves.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>