attr,has: Add Thing.FindAttr + Thing.FindAttrs methods
This change adds two methods to the Thing type and interface. Both
methods use the new Attribute.Is method to find specific Attribute of
the Thing. FindAttr returns the first match found. FindAttrs returns a
slice of all matches or a nil []has.Attribute slice.

These methods allow attributes to be found by directly iterating over
the Thing.attrs slice instead of the current Attribute finders which
call Thing.Attrs to obtain a copy of the Thing.attrs slice.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>