event,attr: Replace Thing.Attrs usage with attr.FindName
The event package is the last remaining user of Thing.Attrs - which is
inefficient as it makes a copy of the Thing.attrs slice. Looping over
the result of Thing.Attrs and manually finding the Name attribute was
done in the first place due to a cyclic import when importing the attrs
package.

To get around the cyclic import we can use the same butt-ugly hack as
used for the event.Script reference. As the attr package can import the
event package we can add an init function to attr/name.go and have it
populate an event.FindName function we can then call later.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>