attr: Always send prompt to reset input colour
At the moment if the prompt style is set to StyleNone nothing is sent
back to the client for the prompt. This results in the input text being
the colour of the last text sent.

This change always sends the prompt, event if the prompt itself is
empty, so that the correct colour is used.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>