attr,has: New Health attribute
Health implements an attribute representing either the state of health
of a Thing or health bonuses/penalities to be applied to a Thing. If the
Health attribute is on a Player or Mobile then the values are absolute
and represent current and maximum health levels of the Player or Mobile.
If the Health attribute is on anything else the values are relative
modifiers applied to a Player or Mobile when the Thing is used, worn,
wielded, eaten, drunk or otherwise applied.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>