frontend: Use Locate.Where as exiting signal in frontend.game
At the moment frontend.game checks the return value of cmd.Parse to see
if the player has issued a QUIT command. If a QUIT command was issued
the player is moved from frontend.game to frontend.menu.

This works as long as it is the player issuing the QUIT command. However
if the player is scripted to QUIT - for example because they die - this
cannot be communicated back to the frontend.

This change uses the fact that the player will be removed from the
world, that is Locate.Where will return nil, when they quit. This is now
used as a signal by frontend.game to return to frontend.menu.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>