cmd: Returning parsed command from cmd.Parse now redundant
Currently the cmd.Parse function returns the command that was parsed.
This was used by frontend.game to monitor for a QUIT command to signal
that the frontend should switch to the main menu frontend.menu.

As frontend.game now uses Locate.Where as a signal to switch frontends
cmd.Parse no longer needs to return the parsed command.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>