cmd: Add state.asParticipant method
This change adds a state.asParticipant method that enables a participant
to execute a sub-command as part of a command. For example a GIVE
command could be implemented by having the actor script a DROP command
followed by the participant scripting a GET command.

Note: It is currently assumed the actor and participant are at the same
location - see method comments for details.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>