attr: Improve Player.buildPrompt method
The Player.buildPrompt method now uses Health.Prompt when building the
short and long prompts. This allows the current health to be colour
coded based on the percentage of maximum health. The short prompt now
only display the current health. As a result of the chnge some colour
bleeding issues have aslo been fixed. A bug causing the prompt style
'none' to still display a '>' has also been fixed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>