cmd,docs: Veto HIT command + COMBAT pseudo command for HIT
This change adds checks to the HIT command for vetoing HIT or the COMBAT
pseudo command. The location is checked first, then everything in it.
This allows for safe areas to be set up where combat is not allowed. It
also allows items and mobiles such as peacekeepers to veto fighting.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>