data: Reinstate garden's sheltered area sanctuary status
In the Java version of WolfMUD there were two safe, sheltered areas in
the garden. Reinstate the sanctuary status of the sheltered areas.

NOTE: Once the Health attribute is working properly for entities, other
than on players, it should be possible to increase the healing effect of
these areas by raising the regeneration rate and/or amount in these
peaceful areas.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>