frontend: Really ensure player has at least one point of health
This is a fix for commit 1b4cfddec5b86ce5 'frontend: Auto upgrade Player
with Health attribute on login'.

The previous commit only checked the player has at least one point of
health when they logged in. However the check should be done whenever
the player enters the game. This fix moves the check from frontend/login
to frontend/game.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>