message, attr: Reduce (fix) number of colour resets sent
There is a check in Buffer.Deliver that is supposed to check if the
buffer starts with an ANSI escape sequence - in which case we don't send
a colour reset sequence as the colour is going to change immediately
anyway.

However this is only working for messages to the actor where omitLF is
set to true - as there is a line feed in the stream already from the
player hitting enter. In the case of participants and observers omitLF
is set to false - so all messages have a line feed prefixed to them
causing the check to always fail.

This change updates the check to only prefix the buffer with a colour
reset if the buffer does not start with an ANSI escape sequence or a
line feed followed by an ANSI escape sequence. This is safe as a plain
line feed cannot reset or alter colours.

Due to this change it is not longer necessary to always send a colour
reset sequence as part of the player prompt.

NOTE: At the moment WolfMUD does not make use of background colours,
although the support is there and working and can be invoked by using
the '[BG<color>]' sequence in data files, for example '[BGBLACK]'. If
background colour changes are being used the above checking could fail
leading to unintentional colour bleed. This is not a new problem
introduced by this change. A work around is to manually include a
'[RESET]' when using background colours.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>