attr: Don't modify passed []byte in Player.Write
Currently Player.Write modifies the passed []byte to append the current
prompt before writing the data. However Player.Write implements the
io.Writer interface, the documentation of which says:

  Write must not modify the slice data, even temporarily.

This change updates Player.Write to make sure it allocates a copy of the
passed []byte when appending the prompt data, thus preserving the
originally passed []byte.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>