message: Don't copy buffer for each io.Writer in Buffer.Deliver
Currently the Buffer.Deliver method makes sure to pass a copy of the
buffer to each io.Writer in case it is modified. However the io.Writer
interface documentation says:

  Write must not modify the slice data, even temporarily.

Therefore we should not have to worry about the []byte passed to Write
being modified.

This change therefore simplifies Buffer.Deliver by not copying the
buffers passed to each Writer.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>