recordjar: Make hyphen/minus '-', underscore '_' not name/value separator
Currently the separator between name/value pairs is detected by finding
the first non-letter, non-digit or non-whitespace rune. Specified names
can be chosen carefully so that they are single words. For user
specified names it is desirable to be able to separate multiple words in
a name - especially as names are case insensitive and will be encoded as
all uppercase.

An example is the, soon to be added, Body attribute where users can
specify the name of body slots with the name of the slot and a quantity:

  Body: Head Eye→2 Ear→2 Nose Mouth UpperLip LowerLip

The recordjar encoder would save this as:

  Body: HEAD EYE→2 EAR→2 NOSE MOUTH UPPERLIP LOWERLIP

If the user tried to specify the last two names as 'Upper_Lip' and
'Lower_Lip' the underscore would be taken to be a name/value separator
resulting in 'UPPER' with a value of 'LIP' and 'LOWER' with a value of
'LIP'.

This change allows a hyphen/minus '-' or underscore '_' to be used as
part of a name and are no longer identified as a name/value separator.
This would allow for the last two names to be specified as either
'Upper-Lip', 'Lower-Lip' or 'Upper_Lip', 'Lower_Lip'.

This change effects the PairList and KeyedStringList decoders and
comments have been updated. The documentation for these types in
docs/zone-files.txt has been updated.

The logic for testing if a rune is a separator has moved from
indexSeparator to a new IsSeparator function as the checking is now too
verbose to be inlined. It also allows complex attr.Unmarshal methods,
that need their own processing, to access the function if needed instead
of rolling their own version.

Tests and benchmarks updated with new cases.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>