recordjar: Allow leading exclamation mark '!' in decoder.PairList
Currently an exclamation mark '!' used in a PairList would be identified
as a keyword/value separator - unless it appeared after the first
detected separator where it would be part of the value.

It is common in computing to use the exclamation mark to mean 'not' or
to invert the truthfulness of something. With the upcoming Body
attribute changes there is a need to specify that a body part is
missing, i.e. specify there is not a body part:

  Body: HEAD EYE !EYE EAR→2 MOUTH

In this example we have two eyes but have lost one 'EYE !EYE'.

This change updates decode.PairList to allow a keyword to be prefixed
with an exclamation mark. This may be extended to other types at a later
time. A keyword prefixed with an exclamation mark is distinct from one
without. In the example 'EYE' and '!EYE' are separate keywords.

New tests have been added for this functionality. The zone file
documentation in docs/zone-files.txt has also been updated.

No changes are required for encode.PairList as it does not depend on
detecting separators.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>