attr,has: New Body attribute + supporting interfaces
This change adds a new Body attribute that is used to define the body
slots available for holding, wearing and wielding items. As well as the
has.Body interface, and it's default implementation attr.Body, there is
a new has.Slottable interface - however this is not a full attribute but
one that should be implemented by attributes needing body slots - such
as Holdable, Wearable and Wieldable.

The interface for Holdable, Wearable and Wieldable have been added but
the default implementations will be added in subsequent commits.

In player and zone files the Body attribute is represented as a PairList
consisting of the name of the slot and optionally the number of slots,
which defaults to one if not specified. For example:

  Body: ANKLE→2 BACK→1 CHEST→1 EAR→2 ELBOW→2 EYE→2 FACE→1 FINGER→8
        FOOT→2 HAND→2 HEAD→1 KNEE→2 LOWER_ARM→2 LOWER_LEG→2 LOWER_LIP→1
        MOUTH→1 NECK→1 NOSE→1 PELVIS→1 SHOULDER→2 THUMB→2 UPPER_ARM→2
        UPPER_LEG→2 UPPER_LIP→1 WAIST→1 WRIST→2

The docs/zone-files.txt has been updated with details of the new Body
attribute.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>