cmd: Sync Body slots in QUIT command
Currently the QUIT command will forcibly junk unique and non-collectable
items in a player's inventory. It may be that such an item is in use and
being held, worn or wielded.

This change forcibly removes any used unique and non-collectable items
before they are forcibly junked so that the body slots are correctly
updated.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>