cmd: Display item usage for INVENTORY command + update text
When listing the items in the player's inventory some of the items may
be in use - held, worn or wielded.

This change adds the current usage of an item to the inventory list.
Current usage appears after the item's name and is highlighted green.
The message 'You are currently carrying' has been updated to 'You
currently have' now items can be held and this could be confusing.

  >inv
  You currently have:
    a leather jerkin  - wearing
    a small green ball
    a dagger - wielded
    a large rock

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>