cmd: Add WIELD command
This change adds a WIELD command allowing items to be wielded and used
as weapons. For an item to be wieldable it must be defined with a
WIELDABLE attribute that specifies the BODY slots required to wield the
item. To successfully wield the item the required BODY slots must be
available and free of other items.

NOTE: Additional work is required in order to have functioning weapons
that actually do damage.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>