cmd: Add WEAR command
This change adds a WEAR command allowing items to be worn. For an item
to be wearable it must be defined with a WEARABLE attribute that
specifies the BODY slots required to wear the item. To successfully wear
the item the required BODY slots must be available and free of other
items.

NOTE: Additional work is required in order to have functionality such as
armour to prevent or reduce damage.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>