cmd: Add HOLD command
This change adds a HOLD command allowing items to be held. By default
all items that are not players or mobiles, that don't have a BODY
attribute, have an implicit HOLDABLE attribute allowing the item to be
held in one hand:

  HOLDABLE: HAND

A specific HOLDABLE attribute can be added to override the default. For
example larger items may require two hands in order to hold them:

  HOLDABLE: HAND→2

A specific HOLDABLE attribute can also be defined for mobiles, maybe so
that small creatures can be held. In this case the HOLDABLE could be for
one hand or two depending on the size of the mobile.

If an item should not be held at all then the HOLD command may be vetoed
for the item:

  VETO: HOLD→You cannot hold that in your hand.

This will override the implicit HOLDABLE attribute.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>