docs: Rewrite Alias/Aliases section in zone-files.txt
The current documentation for the Aliases/Aliases section is inadequate
and just repeats much of the information found in the Barrier section.
It also fails to document qualifiers and bound qualifiers.

This commit rewrites the Alias/Aliases section to deal specifically with
aliases and adds sub-sections for qualified aliases and bound aliases.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>