data: Cleanup chest, pouch, ball reference + move to trading post
This commit cleans up the Refs on the chest, leather pouch and small red
ball so they are sequential again.

While these items were intended just for testing they are some of the
earliest items created and it seems a shame to just remove them :( For
now just relocate them to the trading post.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>