data: Add Merkle's Fine Emporium (clothier, L73)
This change adds Merkle's Fine Emporium to the main Zinara zone as
location L73. It can be found west of Fountain Square. It also adds
Merkle himself and a range of wearable items.

These items are so that players have something to try out the new Body
attribute system. They also act as examples for authors creating their
own items.

Currently items cannot be bought and sold. For now Merkle is happy for
anyone to just take the items available :)

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>